pyproject.toml
[project]
name = "noticias"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.14.0",
    "playwright>=1.57.0",
    "python-dotenv>=1.2.1",
]
